requirements.txt
streamlit>=1.30.0
yfinance>=0.2.36
lxml>=4.9.0
pandas>=2.0.0
plotly>=5.18.0
openai>=1.10.0
textblob>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
